% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{get_hinv_postvar}
\alias{get_hinv_postvar}
\title{Empirical posterior variance of the unconstrained draws}
\usage{
get_hinv_postvar(fit)
}
\arguments{
\item{fit}{a \code{CmdStanMCMC} object.}
}
\value{
a \verb{d x d} covariance matrix of the unconstrained draws.
}
\description{
Empirical posterior variance of the unconstrained draws
}
\keyword{internal}
